Crate namada_replay_protection

Replay protection storage keys

Structs

Key

A storage key is made of storage key segments DbKeySeg, separated by KEY_SEGMENT_SEPARATOR.

Functions

commitment_key

Get the key under which we store a hash which is commitment to all replay protection entries.

current_key

Get the transaction hash key under the current subkey

current_prefix

Get the transaction hash prefix under the current subkey

key

Get the transaction hash key